Refactor out shard planning
